speech-processing-server/requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
Django==3.2.12
djangorestframework==3.13.1
google-api-core==2.10.1
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-speech==2.15.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.49.0
grpcio-status==1.49.0
gunicorn==20.1.0
idna==3.4
jarowinkler==1.2.1
joblib==1.1.0
Levenshtein==0.20.3
numpy==1.23.3
proto-plus==1.22.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.2.1
rapidfuzz==2.10.0
requests==2.28.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
urllib3==1.26.12
